[Added] Authentication Middleware

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import assertIsDefined from "../utils/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find();
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId });
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,8 +19,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const { noteId } = req.params;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id");
     }
@@ -24,6 +32,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You not authorized to access this note");
     }
 
     res.status(200).json(note);
@@ -43,13 +55,20 @@ export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknow
   next
 ) => {
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note title is required");
     }
 
-    const newNote = await NoteModel.create({ title: title, text: text });
+    const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
+      title: title,
+      text: text,
+    });
     res.status(201).json(newNote);
   } catch (error) {
     next(error);
@@ -74,8 +93,11 @@ export const updateNote: RequestHandler<
   const { noteId } = req.params;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id");
     }
@@ -90,6 +112,10 @@ export const updateNote: RequestHandler<
       throw createHttpError(404, "Note not found");
     }
 
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You not authorized to edit this note");
+    }
+
     note.title = newTitle;
     note.text = newText;
 
@@ -102,8 +128,11 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const { noteId } = req.params;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id");
     }
@@ -112,6 +141,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You not authorized to delete this note");
     }
 
     await note.deleteOne();
